fixup! [#1164] added post message communication
