+++ common/src/components/Event/EventFormFields.tsx
@@ -35,7 +35,7 @@ import LocationField from './fields/LocationField'
 import { TitleField } from './fields/TitleField'
 import { VideoConferenceField } from './fields/VideoConferenceField'
 import { TdriveButton } from '@common/features/Tdrive/components/TdriveButton'
-import { TdriveFile } from '@common/features/Tdrive/hooks/useTdrivePicker'
+import { TdriveFile } from '@common/features/Tdrive/types'
 import { Attachment } from '@common/types/Attachment'
 import { useEventFormValues } from './hooks/useEventFormValues'
 import { validateEventFormValues } from './utils/formValidation'
