+++ common/src/features/Tdrive/components/TdriveButton.tsx
@@ -5,7 +5,8 @@ import { AttachmentField } from '@common/components/Event/fields/AttachmentField
 import { Box, Button, Snackbar, Alert } from '@linagora/twake-mui'
 import { useI18n } from 'twake-i18n'
 import { useScreenSizeDetection } from '@common/useScreenSizeDetection'
-import { useTdrivePicker, TdriveFile } from '../hooks/useTdrivePicker'
+import { useTdrivePicker } from '../hooks/useTdrivePicker'
+import { TdriveFile } from '../types'
 import { TdrivePickerDialog } from './TdrivePickerDialog'
 import { Icon, Drive } from '@linagora/twake-icons'
 import { Attachment } from '@common/types/Attachment'
