+++ common/src/features/Tdrive/index.ts
@@ -1,9 +1,7 @@
 export { TdriveButton } from './components/TdriveButton'
 export { TdrivePickerDialog } from './components/TdrivePickerDialog'
 export { useTdrivePicker } from './hooks/useTdrivePicker'
-export type { TdriveFile } from './hooks/useTdrivePicker'
-export { createMockCozyClient } from './cozyClientMock'
-export type { MockCozyClient, MockCozyClientOptions } from './cozyClientMock'
+export type { TdriveFile } from './types'
 export { exchangeToken, createIntent } from './TdriveDao'
 export type {
   TokenExchangeResponse,
