+++ common/src/features/Tdrive/types.ts
@@ -1,0 +1,7 @@
+export interface TdriveFile {
+  id: string
+  name: string
+  url: string
+  type: 'sharingLink' | 'downloadLink'
+  mimeType: string | null
+}
